fix a compilation failure

+++ trunk/linbox/linbox/ffpack/ffpack_charpoly.inl
@@ -54,9 +54,8 @@ FFPACK::CharPoly (const Field& F, std::list<Polynomial>& charp, const size_t N,
 		FFLAS_INT_TYPE p;
 		F.characteristic(p);
 		// Heuristic condition (the pessimistic theoretical one being p<2n^2.
-		if (p < N)
+		if (p < (unsigned long) N)
 			return CharPoly (F, charp, N, A, lda, FfpackLUK);
-
 		do{
 			try {
 				CharpolyArithProg (F, charp, N, A, lda, __FFPACK_CHARPOLY_THRESHOLD);
